#5 — migrate reviewer-opencode tests to airframe-stub pattern
Deletes tests/unit/actors/xoscar_runtime/test_reviewer_opencode.py
(444 LOC, 7 tests) and replaces with test_reviewer_inbox.py (282 LOC,
6 tests). Same supervisor-contract coverage:

* send_review success → typed correctness/completeness_review_ready
  + back-compat review_ready fan-out
* findings round-trip with Pydantic typing intact
* RuntimeAuthError → prompt_error with bead_id attribution
* new_bead → agent.rotate_session()
* send_aggregate_review success → aggregate_review_ready
* send_aggregate_review failure → payload_parse_error (non-fatal)

Dropped from the migration: prompt-content assertions ("Review
context" / "previous review") and OpenCode-session-ID rotation
tests. Those exercise the agent's prompt-building + OpenCode HTTP
session management internals — neither is a concern of the actor
shell. The agent-level prompt tests stay in tests/unit/agents/ and
will be addressed in the agent test migration round.

3568 unit tests pass.
